refactor newsletter api routes
